scripts/pipeline/requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
